requirements.txt
# Core dependencies for Phase 1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Video processing
opencv-python==4.8.1.78
ffmpeg-python==0.2.0
Pillow==10.1.0

# Audio/Speech processing
openai-whisper==20231117
youtube-transcript-api==0.6.1
google-api-python-client==2.110.0  # YouTube Data API for video search
isodate==0.6.1  # For parsing ISO 8601 duration format
yt-dlp>=2024.12.13  # Latest version for better YouTube support

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
pydantic>=2.6.0,<2.10.0
httpx==0.25.2
pathlib==1.0.1
click>=8.0.2,<8.2.0
filelock~=3.16.1
psutil~=6.1.0

# Background processing
celery==5.3.4
redis==5.0.1

# Phase 2: Vector Embeddings & RAG
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.1
torchvision==0.16.1
lancedb==0.3.4
chromadb==0.4.15
langchain==0.0.340
langchain-openai==0.0.2
numpy==1.24.4
scikit-learn==1.3.2

# OpenAI for GPT models
openai==1.3.7

# Phase 3: Conversational Interface
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1

# Phase 4: Visual Search Engine  
ultralytics==8.0.196
detectron2 @ git+https://github.com/facebookresearch/detectron2.git
clip-by-openai==1.0

# Phase 5: Navigation & User Interface
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Additional ML/AI dependencies
accelerate==0.25.0
datasets==2.15.0
evaluate==0.4.1

# WebSocket support for real-time features
python-socketio==5.10.0
websockets==12.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-asyncio==0.21.1
black==23.11.0
